Fixes for vision/data.py
valid_pct was assigned to be of integer type multiple times. Changed it to type float.

+++ fastai/vision/data.py
@@ -327,7 +327,7 @@ class ImageDataBunch(DataBunch):
                 fn_col=fn_col, label_col=label_col, suffix=suffix, header=header, **kwargs)
 
     @classmethod
-    def from_lists(cls, path:PathOrStr, fnames:FilePathList, labels:Collection[str], valid_pct:int=0.2, test:str=None, **kwargs):
+    def from_lists(cls, path:PathOrStr, fnames:FilePathList, labels:Collection[str], valid_pct:float=0.2, test:str=None, **kwargs):
         classes = uniqueify(labels)
         train,valid = random_split(valid_pct, fnames, labels)
         datasets = [ImageClassificationDataset(*train, classes),
@@ -336,12 +336,12 @@ class ImageDataBunch(DataBunch):
         return cls.create(*datasets, path=path, **kwargs)
 
     @classmethod
-    def from_name_func(cls, path:PathOrStr, fnames:FilePathList, label_func:Callable, valid_pct:int=0.2, test:str=None, **kwargs):
+    def from_name_func(cls, path:PathOrStr, fnames:FilePathList, label_func:Callable, valid_pct:float=0.2, test:str=None, **kwargs):
         labels = [label_func(o) for o in fnames]
         return cls.from_lists(path, fnames, labels, valid_pct=valid_pct, test=test, **kwargs)
 
     @classmethod
-    def from_name_re(cls, path:PathOrStr, fnames:FilePathList, pat:str, valid_pct:int=0.2, test:str=None, **kwargs):
+    def from_name_re(cls, path:PathOrStr, fnames:FilePathList, pat:str, valid_pct:float=0.2, test:str=None, **kwargs):
         pat = re.compile(pat)
         def _get_label(fn): return pat.search(str(fn)).group(1)
         return cls.from_name_func(path, fnames, _get_label, valid_pct=valid_pct, test=test, **kwargs)
